requirements.txt
# Python requirements

# Testing
pytest >= 7.1
pytest-xprocess >= 0.18
pytest-order >= 1.0.1
lxml >= 4.9.1

# VISA framework
pyvisa >= 1.11
pyvisa-py >= 0.5
